Addition of counters of all transactions on the board table as well as the list of users who have made a reservation

+++ lib/models/position_model.dart
@@ -2,8 +2,8 @@
 
 class Position {
 
-  int? lat;
-  int? long;
+  double? lat;
+  double? long;
 
   Position({
     this.lat,
